refactor colorsets
