+++ extension/src/prototypes/datasets/colorSets/constants.ts
@@ -1,0 +1,6 @@
+export const LANDSLIDE_RISK_TYPE_CODES = [
+  1, // 警戒区域
+  2, // 特別警戒区域
+  3, // 警戒区域(指定前)
+  4, // 特別警戒区域(指定前)
+];
