+++ extension/src/prototypes/datasets/colorSets/index.ts
@@ -1,3 +1,9 @@
 export * from "./floodRankColorSet";
-export * from "./landSlideRiskColorSet";
 export * from "./landUseColorSet";
+
+export * from "./usageColorSet";
+export * from "./structureTypeColorSet";
+export * from "./fireproofStructureTypeColorSet";
+export * from "./steepSlopeRiskColorSet";
+export * from "./landslideRiskColorSet";
+export * from "./mudflowRiskColorSet";
