--- extension/src/shared/plateau/colorSets/index.ts
@@ -1,1 +1,0 @@
-export * from "./buildingColor";
