+++ extension/src/shared/plateau/index.ts
@@ -1,3 +1,2 @@
 export * from "./layers";
 export * from "./featureInspector";
-export * from "./colorSets";
