+++ extension/src/shared/plateau/layers/TilesetProperties.ts
@@ -3,18 +3,16 @@ import {
   QualitativeColor,
   QualitativeColorSet,
   atomsWithQualitativeColorSet,
-} from "../../../prototypes/datasets";
-import { isNotNullish } from "../../../prototypes/type-helpers";
-import { Properties } from "../../reearth/utils";
-import { TilesetFloodColorField } from "../../types/fieldComponents/3dtiles";
-import {
   usageColorSet,
   structureTypeColorSet,
   fireproofStructureTypeColorSet,
   steepSlopeRiskColorSet,
   mudflowRiskColorSet,
-  landSlideRiskColorSet,
-} from "../colorSets";
+  landslideRiskColorSet,
+} from "../../../prototypes/datasets";
+import { isNotNullish } from "../../../prototypes/type-helpers";
+import { Properties } from "../../reearth/utils";
+import { TilesetFloodColorField } from "../../types/fieldComponents/3dtiles";
 import { makePropertyName } from "../featureInspector";
 
 export type AvailableFeatures = ("color" | "buildingFilter" | "floodFilter")[];
@@ -92,7 +90,7 @@ const qualitativeProperties: QualitativeProperty[] = [
   },
   {
     testProperty: propertyName => propertyName === "土砂災害リスク_地すべり_区域区分コード",
-    colorSet: landSlideRiskColorSet,
+    colorSet: landslideRiskColorSet,
     getDisplayName: () => "地すべり",
     availableFeatures: ["color"],
   },
